requirements.txt
google-cloud-logging
google-generativeai
requests
dotenv
sqlalchemy 
psycopg2-binary
fastapi
streamlit
plotly
pandas
networkx
scikit-learn
numpy
python-dateutil
streamlit-chat
streamlit-option-menu
streamlit-aggrid